+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [next]
 
 - chore(): cleanup Type assertions [#8950](https://github.com/fabricjs/fabric.js/pull/8950)
+- BREAKING: fabric.util.makeElementSelectable / fabric.util.makeElementUnselectable are removed [#8930](https://github.com/fabricjs/fabric.js/pull/8930)
+- refactor(): Canvas DOM delegation to utility class [#8930](https://github.com/fabricjs/fabric.js/pull/8930)
 
 ## [6.0.0-beta7]